Separate portal from shared core entry
