+++ vuu-ui/packages/core/src/index.ts
@@ -4,14 +4,3 @@ export {
   DataProvider,
   useData,
 } from "./context-definitions/DataProvider";
-export {
-  PortalHeader,
-  type PortalHeaderProps,
-} from "./portal-header/PortalHeader";
-export { PortalNav, type PortalNavProps } from "./portal-nav/PortalNav";
-export { PortalShell, type PortalShellProps } from "./portal-shell/PortalShell";
-export {
-  RemoteModule,
-  type RemoteModuleProps,
-} from "./remote-module/RemoteModule";
-export type { RemoteModuleDescriptor } from "./RemoteModuleDescriptor";
