+++ vuu-ui/packages/core/src/portal.ts
@@ -1,0 +1,11 @@
+export {
+  PortalHeader,
+  type PortalHeaderProps,
+} from "./portal-header/PortalHeader";
+export { PortalNav, type PortalNavProps } from "./portal-nav/PortalNav";
+export { PortalShell, type PortalShellProps } from "./portal-shell/PortalShell";
+export {
+  RemoteModule,
+  type RemoteModuleProps,
+} from "./remote-module/RemoteModule";
+export type { RemoteModuleDescriptor } from "./RemoteModuleDescriptor";
