+++ vuu-ui/packages/vuu-shell/src/feature/Feature.tsx
@@ -1,7 +1,7 @@
-import { RemoteModule } from "@vuu-ui/core";
+import { RemoteModule } from "@vuu-ui/core/portal";
 import { registerComponent } from "@vuu-ui/vuu-utils";
 
-/** @deprecated Use RemoteModule from @vuu-ui/core. */
+/** @deprecated Use RemoteModule from @vuu-ui/core/portal. */
 export const Feature = RemoteModule;
 Feature.displayName = "Feature";
 registerComponent("Feature", Feature, "view");
