+++ vuu-ui/packages/vuu-shell/src/feature/index.ts
@@ -2,4 +2,4 @@ export * from "./Feature";
 export {
   RemoteModule,
   type RemoteModuleProps,
-} from "@vuu-ui/core";
+} from "@vuu-ui/core/portal";
